move Bytes and DataInit to a separate module
